requirements.txt
# Production dependencies for bedtime-story API
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
pydantic-settings>=2.0,<3
openai>=1.0,<2
python-dotenv>=1.0,<2
python-jose[cryptography]>=3.3,<4
